delete button includes double click & color confirmation

+++ app/src/main/java/com/example/mynext/ui/ItemDialogBuilder.kt
@@ -5,22 +5,26 @@ import android.app.Dialog
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.util.Log
 import android.view.View
 import com.example.mynext.R
 import com.example.mynext.fragments.ItemDialogCallback
 import com.example.mynext.model.Item
 import com.example.mynext.util.ImageHelper
+import com.google.android.material.button.MaterialButton
 import kotlinx.android.synthetic.main.dialog_item_details.view.*
 import java.text.DateFormat
 import java.util.*
 
-class ItemDialogBuilder (val context: Context, private val itemDialogCallback: ItemDialogCallback) {
+class ItemDialogBuilder(val context: Context, private val itemDialogCallback: ItemDialogCallback) {
 
-    fun getItemDialog(item: Item, itemsName: String?) : Dialog {
+    fun getItemDialog(item: Item, itemsName: String?): Dialog {
 
         val dialogBuilder = AlertDialog.Builder(context)
         val dialogView = View.inflate(context, R.layout.dialog_item_details, null)
+
+        dialogView.setOnClickListener {
+            resetDeleteButton(dialogView.itemdialog_delete_btn)
+        }
 
         setItemInformation(dialogView, item)
 
@@ -50,10 +54,20 @@ class ItemDialogBuilder (val context: Context, private val itemDialogCallback: I
             itemdialog_image_iv.setImageBitmap(bitmap)
 
             if (item.done) {
-                itemdialog_markasdone_btn.setBackgroundColor(resources.getColor(R.color.colorDone, context.theme))
+                itemdialog_markasdone_btn.setBackgroundColor(
+                    resources.getColor(
+                        R.color.colorDone,
+                        context.theme
+                    )
+                )
                 itemdialog_markasdone_btn.setIconResource(R.drawable.ic_baseline_check_box_24)
             } else {
-                itemdialog_markasdone_btn.setBackgroundColor(resources.getColor(R.color.colorAccent, context.theme))
+                itemdialog_markasdone_btn.setBackgroundColor(
+                    resources.getColor(
+                        R.color.colorAccent,
+                        context.theme
+                    )
+                )
                 itemdialog_markasdone_btn.setIconResource(R.drawable.ic_baseline_check_box_outline_blank_24)
             }
         }
@@ -61,34 +75,60 @@ class ItemDialogBuilder (val context: Context, private val itemDialogCallback: I
 
     private fun setButtonsActions(dialogView: View, item: Item, itemsName: String, dialog: Dialog) {
 
-        with(dialogView) {
-            itemdialog_findonline_btn.setOnClickListener {
+        with(dialogView.itemdialog_findonline_btn) {
+            setOnClickListener {
                 //Example of generated url: https://www.google.com/search?q=Book+To+kill+a+Mockingbird
                 val url = "https://www.google.com/search?q=$itemsName+${item.itemTitle}"
                 val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
                 context.startActivity(browserIntent)
-            }
 
-            itemdialog_markasdone_btn.setOnClickListener {
+                resetDeleteButton(dialogView.itemdialog_delete_btn)
+            }
+        }
+
+        with(dialogView.itemdialog_markasdone_btn) {
+            setOnClickListener {
                 if (item.done) {
                     item.done = false
-                    itemdialog_markasdone_btn.setBackgroundColor(resources.getColor(R.color.colorAccent, context.theme))
-                    itemdialog_markasdone_btn.setIconResource(R.drawable.ic_baseline_check_box_outline_blank_24)
+                    setBackgroundColor(resources.getColor(R.color.colorAccent, context.theme))
+                    setIconResource(R.drawable.ic_baseline_check_box_outline_blank_24)
                 } else {
                     item.done = true
-                    itemdialog_markasdone_btn.setBackgroundColor(resources.getColor(R.color.colorDone, context.theme))
-                    itemdialog_markasdone_btn.setIconResource(R.drawable.ic_baseline_check_box_24)
+                    setBackgroundColor(resources.getColor(R.color.colorDone, context.theme))
+                    setIconResource(R.drawable.ic_baseline_check_box_24)
                 }
                 itemDialogCallback.updateItem(item)
-            }
 
-            itemdialog_edit_btn.setOnClickListener {
-                Log.d("MYTAG", "edit item")
+                resetDeleteButton(dialogView.itemdialog_delete_btn)
             }
+        }
 
-            itemdialog_delete_btn.setOnClickListener {
-                itemDialogCallback.deleteItem(item)
-                dialog.dismiss()
+        with(dialogView.itemdialog_edit_btn) {
+            setOnClickListener {
+                //TODO Implement edit action
+                resetDeleteButton(dialogView.itemdialog_delete_btn)
+            }
+        }
+
+
+        with(dialogView.itemdialog_delete_btn) {
+            setOnClickListener {
+                if (tag == null) {
+                    tag = "delete"
+                    setBackgroundColor(resources.getColor(R.color.colorError, context.theme))
+                } else {
+                    itemDialogCallback.deleteItem(item)
+                    dialog.dismiss()
+                }
+            }
+        }
+    }
+
+    private fun resetDeleteButton(button: MaterialButton) {
+        with(button) {
+            if (tag != null) {
+                tag = null
+                setBackgroundColor(resources.getColor(R.color.colorAccent, context.theme))
             }
         }
     }
